Added Graph class, Dijkstra algorithm segmenting
Added Graph class and Dijkstra's Algorithm method to ATB. Novel segmentation method that treats each potential segment as a node on the graph, with an edge weight that decreases the longer the word is. Using Dijkstra's algo we transit the graph from start to finish through the least weighted path to give us the segmented sentence, biased toward longest words.